springboard exercise
index.html, app.js, and pup-in-cup.jpg are all a part of the same practice; however, movie.html and movie.js are a sub-part of the jquery exercise. The movie app helped me to understand how jQuery interacts with the DOM and the HTML. Still need to work out a few kinks with the app functionality, but it's a start.